requirements.txt
email==6.0.0a1
numpy==1.17.4
opencv-python==3.4.6.27
pyzmq==18.1.1
RPi.GPIO==0.7.0
